feat: modify skeleton
