--- src/llgen/llgen.h
@@ -1,6 +1,0 @@
-#ifndef _LLGEN_H_
-#define _LLGEN_H_
-
-void run(FILE *in, FILE* out);
-
-#endif
