--- src/llgen/tests/testin1.txt
@@ -1,25 +1,0 @@
-[terminal]
-
-+ - * /
-( )
-num name
-eof
-epsilon
-
-[nonterminal]
-
-goal
-expr
-expr'
-term
-term'
-factor
-
-[production]
-
-goal -> expr $
-expr -> term expr' $
-expr' -> + term expr' | - term expr' | epsilon $
-term -> term * factor | term / factor | factor $
-term' -> * factor term' | / factor term' | epsilon $
-factor -> ( expr ) | num | name $
