+++ ui/lib/darray.c
