+++ ui/lib/stack.c
