+++ ui/lib/stack.h
