+++ ui/toy/scanner.c
