+++ ui/toy/scanner.h
