+++ ui/toy/symtable.c
