+++ ui/toy/symtable.h
